+++ packages/crc32/src/index.ts
@@ -1,3 +1,5 @@
+import {uint32ArrayFrom} from "@aws-crypto/util";
+
 export function crc32(data: Uint8Array): number {
   return new Crc32().update(data).digest();
 }
@@ -20,7 +22,7 @@ export class Crc32 {
 }
 
 // prettier-ignore
-const lookupTable = Uint32Array.from([
+const a_lookUpTable = [
   0x00000000, 0x77073096, 0xEE0E612C, 0x990951BA,
   0x076DC419, 0x706AF48F, 0xE963A535, 0x9E6495A3,
   0x0EDB8832, 0x79DCB8A4, 0xE0D5E91E, 0x97D2D988,
@@ -85,4 +87,6 @@ const lookupTable = Uint32Array.from([
   0xBAD03605, 0xCDD70693, 0x54DE5729, 0x23D967BF,
   0xB3667A2E, 0xC4614AB8, 0x5D681B02, 0x2A6F2B94,
   0xB40BBE37, 0xC30C8EA1, 0x5A05DF1B, 0x2D02EF8D,
-]);
+];
+const lookupTable: Uint32Array = uint32ArrayFrom(a_lookUpTable)
+export { AwsCrc32 } from "./aws_crc32";